+++ hana/lib/HANAService.js
@@ -64,12 +64,7 @@ class HANAService extends SQLService {
   }
 
   async set(variables) {
-    const columns = Object.keys(variables).map(
-      k => `SET '${k.replace(/'/g, "''")}'='${(variables[k] + '').replace(/'/g, "''")}';`,
-    )
-    const sql = `DO BEGIN ${columns.join('')} END;`
-
-    await this.dbc.exec(sql)
+    this.dbc.set(variables)
   }
 
   async onSELECT({ query, data }) {
@@ -83,7 +78,10 @@ class HANAService extends SQLService {
     if (rows.length) {
       rows = this.parseRows(rows)
     }
-    if (cqn.SELECT.count) rows.$count = await this.count(query, rows)
+    if (cqn.SELECT.count) {
+      // REVISIT: the runtime always expects that the count is preserved with .map, required for renaming in mocks
+      return HANAService._arrayWithCount(rows, await this.count(query, rows))
+    }
     return cqn.SELECT.one || query.SELECT.from.ref?.[0].cardinality?.max === 1 ? rows[0] || null : rows
   }
 
@@ -141,7 +139,9 @@ class HANAService extends SQLService {
       })
       .join(' UNION ALL ')
 
-    const ret = `DO BEGIN ${values} SELECT * FROM (${unions}) ORDER BY "_path_" ASC; END;`
+    const ret = temporary.length === 1
+      ? `SELECT _path_ as "_path_",_blobs_ as "_blobs_",_expands_ as "_expands_",_json_ as "_json_"${blobColumns} FROM (SELECT ${temporary[0].select})`
+      : `DO BEGIN ${values} SELECT * FROM (${unions}) ORDER BY "_path_" ASC; END;`
     DEBUG?.(ret)
     return ret
   }
@@ -419,7 +419,7 @@ class HANAService extends SQLService {
         // Making each row a maximum size of 2gb instead of the whole result set to be 2gb
         // Excluding binary columns as they are not supported by FOR JSON and themselves can be 2gb
         const rawJsonColumn = sql.length
-          ? `(SELECT ${sql} FROM DUMMY FOR JSON ('format'='no', 'omitnull'='no', 'arraywrap'='no') RETURNS NCLOB) AS _json_`
+          ? `(SELECT ${sql} FROM DUMMY FOR JSON ('format'='no', 'omitnull'='no', 'arraywrap'='no') RETURNS NVARCHAR(2147483647)) AS _json_`
           : `TO_NCLOB('{}') AS _json_`
 
         let jsonColumn = rawJsonColumn