+++ sqlite/lib/SQLiteService.js
@@ -15,6 +15,9 @@ class SQLiteService extends SQLService {
         dbc.function('session_context', key => dbc[$session][key])
         dbc.function('regexp', { deterministic: true }, (re, x) => (RegExp(re).test(x) ? 1 : 0))
         dbc.function('ISO', { deterministic: true }, d => d && new Date(d).toISOString())
+        dbc.function('json_merge', { varargs: true, deterministic: true }, (...args) =>
+          args.join('').replace(/}{/g, ','),
+        )
         if (!dbc.memory) dbc.pragma('journal_mode = WAL')
         return dbc
       },
@@ -130,8 +133,18 @@ class SQLiteService extends SQLService {
     return this.dbc.exec(sql)
   }
 
-  static CQN2SQL = class CQN2SQLite extends SQLService.CQN2SQL {
+  onPlainSQL({ query, data }, next) {
+    if (typeof query === 'string') {
+      // REVISIT: this is a hack the target of $now might not be a timestamp or date time
+      // Add input converter to CURRENT_TIMESTAMP inside views using $now
+      if (/^CREATE VIEW.* CURRENT_TIMESTAMP[( ]/is.test(query)) {
+        query = query.replace(/CURRENT_TIMESTAMP/gi, "STRFTIME('%Y-%m-%dT%H:%M:%fZ','NOW')")
+      }
+    }
+    return super.onPlainSQL({ query, data }, next)
+  }
 
+  static CQN2SQL = class CQN2SQLite extends SQLService.CQN2SQL {
     column_alias4(x, q) {
       let alias = super.column_alias4(x, q)
       if (alias) return alias
@@ -196,13 +209,7 @@ class SQLiteService extends SQLService {
     }
 
     // Used for SQL function expressions
-    static Functions = { ...super.Functions,
-      // Ensure ISO strings are returned for date/time functions
-      current_timestamp: () => 'ISO(current_timestamp)',
-      // SQLite doesn't support arguments for current_date and current_time
-      current_date: () => 'current_date',
-      current_time: () => 'current_time',
-    }
+    static Functions = { ...super.Functions, ...require('./func') }
 
     // Used for CREATE TABLE statements
     static TypeMap = {
@@ -214,8 +221,12 @@ class SQLiteService extends SQLService {
       Timestamp: () => 'TIMESTAMP_TEXT',
     }
 
-    get is_distinct_from_() { return 'is not' }
-    get is_not_distinct_from_() { return 'is' }
+    get is_distinct_from_() {
+      return 'is not'
+    }
+    get is_not_distinct_from_() {
+      return 'is'
+    }
 
     static ReservedWords = { ...super.ReservedWords, ...require('./ReservedWords.json') }
   }