+++ sqlite/CHANGELOG.md
@@ -1,8 +1,20 @@
-# Change Log
+# Changelog
 
 - All notable changes to this project are documented in this file.
 - The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - This project adheres to [Semantic Versioning](http://semver.org/).
+
+## [1.3.1](https://github.com/cap-js/cds-dbs/compare/v1.3.0...v1.3.1) (2023-10-10)
+
+### Changed
+
+- Updated minimum required version of `@cap-js/db-service`.
+
+## Version 1.3.0 - 2023-10-06
+
+### Fixed
+
+- `CURRENT_TIMESTAMP` in view definition preserves the timezone. #254 
 
 ## Version 1.2.1 - 2023-09-08